Issue: Game checking block underneath player.
Added a line of code adding 0.5 units to the Y coordinate to reduce the likelihood of the game checking the block underneath the player. This would cause the mod to return a light level of 0 if the block is solid, resulting in the player getting stalked, and killed by the Grue in broad daylight.